Generate binlogs for all MSBuild targets
When running an MSBuild target for the project specify the binlog.
This requires a custom build of MonoDevelop and only affects MSBuild
targets that are not being run during a build session.

+++ src/MonoDevelop.ProjectSystem.Tools/MonoDevelop.ProjectSystem.Tools/MSBuildTarget.cs
@@ -114,6 +114,11 @@ namespace MonoDevelop.ProjectSystem.Tools
 			}
 
 			try {
+				if (File.Exists (BinLogFileName)) {
+					copiedBinLogFile = true;
+					return;
+				}
+
 				File.Copy (BuildSessionBinLogFileName, BinLogFileName);
 			} catch (Exception ex) {
 				LoggingService.LogError (
